Updating expanders
