+++ tpgtools/ignored_handwritten/expanders.go
@@ -2,20 +2,6 @@ package google
 
 import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-func expandStringArray(v interface{}) []string {
-	arr, ok := v.([]string)
-
-	if ok {
-		return arr
-	}
-
-	if arr, ok := v.(*schema.Set); ok {
-		return convertStringSet(arr)
-	}
-
-	return convertStringArr(v.([]interface{}))
-}
-
 func expandComputeRouteNextHopInstance(v interface{}, d TerraformResourceData, config *Config) *string {
 	if v == "" {
 		return nil
